why are we on this earth... just to suffer
