--- src/app/home/store/actions.ts
@@ -1,11 +1,0 @@
-import { createAction, props } from "@ngrx/store";
-import { IPost } from "src/app/shared/interfaces/IPost";
-
-const namespace = '[POSTS]'
-
-export const loadPosts = createAction(
-    `${namespace} load posts`,
-    props<{ posts: IPost[] }>(),
-)
-
-
