+++ src/app/store/selectors.ts
@@ -1,5 +1,6 @@
 import { createSelector } from "@ngrx/store";
 import { IGlobalState } from ".";
+import { IPost } from "../shared/interfaces/IPost";
 import { IUserState } from "./userReducer";
 
 export const selectUserState = (state: IGlobalState) => state.user;
@@ -12,4 +13,11 @@ export const selectUser = createSelector(
 export const selectIsAuth = createSelector(
     selectUserState,
     (state: IUserState) => state.isAuth,
+)
+
+export const selectPostsState = (state: IGlobalState) => state.posts;
+
+export const selectAllPosts = createSelector(
+    selectPostsState,
+    (state) => state.all
 )
